pyproject.toml
[project]
name = "sdn-api"
version = "0.1.0"
description = "SDN Watchlist API with two-step matching"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.6",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "httpx>=0.25",
    "black>=23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ['py310']